applying patch to make crawler work with more "recent" realurl versions (not written by us...)
This patch has been around for years but never integrated into any official
release I know of. Not sure where we got it from (I had to retrieve it from
 our internal knowledgebase), but a few years ago it was prominently known
on multiple boards and even on the old TYPO3 bug tracker. The issue can no
longer be found, so I honestly don't know whom to attribute it to, sorry...